+++ index.js
@@ -1,5 +1,0 @@
-import {replaceInFile, replaceInFileSync} from './src/replace-in-file.js'
-import {processFile, processFileSync} from './src/process-file.js'
-
-//Export
-export {replaceInFile, replaceInFileSync, processFile, processFileSync}