+++ src/replace-in-file.spec.ts
@@ -1,11 +1,11 @@
-import {expect, use, should} from 'chai'
+import {expect, use} from 'chai'
 import chaiAsPromised from 'chai-as-promised'
-import {replaceInFile, replaceInFileSync} from './replace-in-file.js'
+import {replaceInFile, replaceInFileSync} from './replace-in-file.ts'
+import type {AsyncFs, SyncFs} from './types.ts'
 import fsAsync from 'node:fs/promises'
 import fs from 'node:fs'
 
-//Enable should assertion style for usage with chai-as-promised
-should()
+//Enable promise assertions
 use(chaiAsPromised)
 
 /**
@@ -193,11 +193,11 @@ describe('Replace in file', () => {
     })
 
     it('should fulfill the promise on success', () => {
-      return replaceInFile({
+      return expect(replaceInFile({
         files: 'test1',
         from: /re\splace/g,
         to: 'b',
-      }).should.be.fulfilled
+      })).to.eventually.be.fulfilled
     })
 
     it('should reject the promise with an error on failure', () => {
@@ -209,12 +209,12 @@ describe('Replace in file', () => {
     })
 
     it('should not reject the promise if allowEmptyPaths is true', () => {
-      return replaceInFile({
+      return expect(replaceInFile({
         files: 'nope',
         allowEmptyPaths: true,
         from: /re\splace/g,
         to: 'b',
-      }).should.be.fulfilled
+      })).to.eventually.be.fulfilled
     })
 
     it('should return a results array', done => {
@@ -365,7 +365,7 @@ describe('Replace in file', () => {
         files: 'test1',
         from: /re\splace/g,
         to: 'b',
-        glob: null,
+        glob: null as any,
       }).then(() => {
         const test1 = fs.readFileSync('test1', 'utf8')
         expect(test1).to.equal('a b c')
@@ -440,9 +440,9 @@ describe('Replace in file', () => {
     describe('fs', () => {
       it('reads and writes using a custom fs when provided', done => {
         const before = 'a'
-        let written
+        let written: string | undefined
 
-        const fs = {
+        const fs: AsyncFs = {
           readFile: async () => {
             return before
           },
@@ -790,15 +790,14 @@ describe('Replace in file', () => {
     describe('fsSync', () => {
       it('reads and writes using a custom fsSync when provided', () => {
         const before = 'a'
-        let written
+        let written: string | undefined
 
-        const fsSync = {
+        const fsSync: SyncFs = {
           readFileSync: () => {
             return before
           },
           writeFileSync: (_fileName, data) => {
             written = data
-            return data
           },
         }